API V2.6. Hay que terminar de comproabr la parte de categoria , una vez comprobada repasara los detalles que he dejado apuntados en el Trello que hay que repasar y darle ya para enganchar la API con Android. Porbar todo RetroFit. Ya coge los datos y los muestra, ahora hay que añadir un producto. Ya se añaden prodcutos mediante el formualrio , pero revisar todo

+++ src/main/java/org/example/entidades/Producto.java
@@ -17,7 +17,7 @@ public class Producto {
 
     @Id
     @Schema(example = "1", description = "Es el identificador del prodcuto, es el código resultante del escaneo")
-    private int codigoBarras;
+    private Long codigoBarras;
 
 
     @Schema(example = "Batido Avena Chocolate 250ml", description = "Nombre del producto")
